Made the .gitignore file less restrictive
I changed the .gitignore here-doc within the script to only ignore
the output and the dataset folder.  The reason I did this was because
I want the project directory to be more complete based on the version
history.  In particular, since I occasionally change the functions files
(e.g. macros.sas), I wanted the version control to capture the state of
the functions so that they match with the analysis scripts.