Major changes, added directories + files, cleaned
To make the copying of the template files easier, I just put the
directories in the template folder so that all I would have to do was to
copy the template folder over. Don't know why I didn't think of this
before. I also added some files in the scripts folder to be more modular
in their activity and editted them down. Cleaned up some of the files,
especially the prodigen script itself.